improve database schema and recipe amount
