+++ db/detailRecipeApi.rb
@@ -3,13 +3,21 @@ require 'net/http'
 require 'openssl'
 require 'json'
 
-types = ["main course", "side dish", "dessert", "appetizer", "salad", "bread", "breakfast", "soup", "beverage", "sauce", "drink"]
-cuisines = ["african", "chinese", "japanese", "korean", "vietnamese", "thai", "indian", "british", "french", "italian","mexican", "spanish", "american", "cajun", "greek", "german"]
-tot_resp = Hash.new
+# types = ["main course", "side dish", "dessert", "appetizer", "salad", "bread", "breakfast", "soup", "beverage", "sauce", "drink"]
+# cuisines = ["african", "chinese", "japanese", "korean", "vietnamese", "thai", "indian", "british", "french", "italian","mexican", "spanish", "american", "cajun", "greek", "german"]
+file = File.read("./rawdata/recipes_id_1111.json")
+resp_json = JSON.parse(file)
+print "Recipe Number: "
+puts resp_json["results"].length()
 
-types.each do |type|
-    puts type
-    url = URI("https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/searchComplex?limitLicense=false&minCalories=0&maxCalories=1500&offset=0&number=100&cuisine="+URI.encode(type))
+tot_resp = Hash.new
+hash_arr = []
+resp_json["results"].each do |recipe|
+    recipe_id = recipe["id"].to_s
+    recipe_kcal = recipe["calories"]
+    puts recipe_id
+
+    url = URI("https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/recipes/"+URI.encode(recipe_id)+"/information")
 
     http = Net::HTTP.new(url.host, url.port)
     http.use_ssl = true
@@ -20,16 +28,17 @@ types.each do |type|
     request["X-RapidAPI-Host"] = 'spoonacular-recipe-food-nutrition-v1.p.rapidapi.com'
 
     response = http.request(request)
-
     resp = JSON.parse(response.read_body)
+    resp["calories"] = recipe_kcal
+
     if tot_resp.empty?
-        tot_resp = resp
+        hash_arr.push(resp)
+        tot_resp = {"recipes"=> hash_arr}
     else
-        tot_resp["results"] = tot_resp["results"] + resp["results"]
+        tot_resp["recipes"] = tot_resp["recipes"].append(resp)
     end
-    
 end
 
-File.open("rawdata/recipes_id_1111.json","w") do |f|
+File.open("rawdata/recipes_1112.json","w") do |f|
     f.write(tot_resp.to_json)
 end
