+++ db/migrate/20221027141304_create_recipes.rb
@@ -6,6 +6,7 @@ class CreateRecipes < ActiveRecord::Migration
       t.string :image_url
       t.integer :origin_id, :null => false, index: { unique: true, name: 'unique_recipe' }
       t.timestamps null: false
+      t.integer :calories
     end
   end
 end
