--- db/rawdata/recipes_id_1110.json
@@ -1,1 +1,0 @@
-{"results":[{"id":482788,"title":"Garlicky Pasta with Swiss Chard and Beans","image":"https://spoonacular.com/recipeImages/482788-312x231.jpg","imageType":"jpg"},{"id":584549,"title":"Stuffed Sweet Potato with Spinach, Hummus & Feta","image":"https://spoonacular.com/recipeImages/584549-312x231.jpg","imageType":"jpg"},{"id":523145,"title":"Nutella\" Overnight Dessert Oats","image":"https://spoonacular.com/recipeImages/523145-312x231.jpg","imageType":"jpg"},{"id":621581,"title":"Endive Stuffed with Old Bay Crab Salad","image":"https://spoonacular.com/recipeImages/621581-312x231.jpg","imageType":"jpg"},{"id":616564,"title":"Chinese Chop Salad + Peanut Sesame Dressing","image":"https://spoonacular.com/recipeImages/616564-312x231.jpg","imageType":"jpg"},{"id":475072,"title":"Spinach-Artichoke French Bread Pizza {Saturdays with Rachael Ray}","image":"https://spoonacular.com/recipeImages/475072-312x231.jpg","imageType":"jpg"},{"id":495020,"title":"Pumpkin Spice Breakfast Shake","image":"https://spoonacular.com/recipeImages/495020-312x231.jpg","imageType":"jpg"},{"id":723373,"title":"Easy Tuscan Bean Soup","image":"https://spoonacular.com/recipeImages/723373-312x231.jpg","imageType":"jpg"},{"id":729431,"title":"Immune Booster Sweet Green Smoothie","image":"https://spoonacular.com/recipeImages/729431-312x231.jpg","imageType":"jpg"},{"id":616091,"title":"Cauliflower in Puttanesca Sauce","image":"https://spoonacular.com/recipeImages/616091-312x231.jpg","imageType":"jpg"},{"id":729431,"title":"Immune Booster Sweet Green Smoothie","image":"https://spoonacular.com/recipeImages/729431-312x231.jpg","imageType":"jpg"}],"baseUri":"https://spoonacular.com/recipeImages/","offset":0,"number":1,"totalResults":113511,"processingTimeMs":264}
