+++ db/schema.rb
@@ -11,9 +11,46 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 0) do
+ActiveRecord::Schema.define(version: 20221027152548) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
+  create_table "ingredients", force: :cascade do |t|
+    t.string   "name",                null: false
+    t.float    "calorie_per_serving"
+    t.datetime "created_at",          null: false
+    t.datetime "updated_at",          null: false
+  end
+
+  add_index "ingredients", ["name"], name: "unique_ingredient", unique: true, using: :btree
+
+  create_table "recipe_ingredients", force: :cascade do |t|
+    t.string   "unit"
+    t.float    "amount",        default: 0.0
+    t.string   "metric_unit"
+    t.float    "metric_amount", default: 0.0
+    t.datetime "created_at",                  null: false
+    t.datetime "updated_at",                  null: false
+    t.integer  "recipe_id"
+    t.integer  "ingredient_id"
+  end
+
+  add_index "recipe_ingredients", ["ingredient_id"], name: "index_recipe_ingredients_on_ingredient_id", using: :btree
+  add_index "recipe_ingredients", ["recipe_id"], name: "index_recipe_ingredients_on_recipe_id", using: :btree
+
+  create_table "recipes", force: :cascade do |t|
+    t.string   "name",       null: false
+    t.text     "steps",      null: false
+    t.string   "image_url"
+    t.integer  "origin_id",  null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.integer  "calories"
+  end
+
+  add_index "recipes", ["origin_id"], name: "unique_recipe", unique: true, using: :btree
+
+  add_foreign_key "recipe_ingredients", "ingredients"
+  add_foreign_key "recipe_ingredients", "recipes"
 end
